requirements.txt
crewai[google-genai]
python-dotenv